Fixed warnings
* fixed lack of use of NSUInteger
* removed deprecated use of triggerChangeNotificationsForDependentKey:

+++ SmugMugCore/SMEAlbumRef.m
@@ -28,7 +28,7 @@
 	return [[[self class] alloc] initWithId:anId key:aKey];
 }
 
--(unsigned int)hash {
+-(NSUInteger)hash {
 	return 31 * [[self albumId] hash] + [[self albumKey] hash];
 }
